requirements.txt
pygame==2.5.2
python-chess==1.999
numpy==1.24.3
torch==2.1.0
tensorboard==2.15.1
